Move large components to containers; create SosModelRunItem
